Add files via upload
Project Aim:

The aim of this project is to visualize the relationship between the 'average weight' and 'highest movement speed' variables using a scatter plot. By extracting these columns from the dataset and plotting them, the project seeks to provide insights into any potential correlation or pattern between weight and speed. The scatter plot serves as a visual representation of the data, allowing for a better understanding of the relationship between these two variables.